resolve small bug

+++ Assets/Scripts/GameLoop.cs
@@ -9,8 +9,8 @@ public class GameLoop : MonoBehaviour
     public float gameDuration;
     public float gameTimer;
     public bool gameIsRunning;
-    public int actualMoney;
-    public int maxMoney;
+    public float actualMoney;
+    public float maxMoney;
     public int[] moneyJalons; //Bronze, Silver, Gold, etc...
     public int commandsSuccesful;
     public int commandsFailed;
